face_attendance/requirements.txt
flask
opencv-python
face-recognition
numpy
pandas
git+https://github.com/ageitgey/face_recognition_models
setuptools
psycopg2-binary
python-dotenv
supabase
